move main code to main
