+++ src/main_lib/mod.rs
@@ -1,0 +1,2 @@
+#[cfg(all(target_os = "windows", feature = "gui"))]
+pub(crate) mod win_gui;
